Added language files for the addon
It is strange that the addon language files are in a completely other format than the basic game files. But all languages are bundled in this file and it has an readable xml format instead of a property txt format. This is only the addon language data; NOT the one from the basic game. 
The German words of the addon can be also found as txt format but it doesn't make sense to use them because there is no english/spanish/... txt file with these information. And here everything is bundled in one readable xml...